Ready for extreme reuse
