+++ ParkingLot.Fotex/Program.cs
@@ -13,19 +13,7 @@ namespace ParkingLot.Fotex
 
             var cli = new ParkingLotCLI(parkingLot);
 
-            while (!cli.Exit)
-            {
-                Console.Write(": ");
-                var input = Console.ReadLine().Split(" ");
-
-                var output = cli.Command(input[0], input.Skip(1).ToArray());
-                if (!string.IsNullOrWhiteSpace(output))
-                {
-                    Console.Write("! ");
-                    Console.WriteLine(output);
-                    Console.WriteLine();
-                }
-            }
+            cli.Run();
         }
     }
 }
